Many major changes (So much recode)
Recoded all the message system to optimize it more and add a lang system, where the configurator chooses the language if it is optional with the plugin.

Also changed the way many variables and some methods are stored, to improve the security and make the access efficient and easy for the plugin using the framework.
With that same fixed errors that happened before, by a bad structure.
Newbie code errors.