Big version with some changes and new feature
This version have the new feature of debug report command, that create a zip file of the plugin files and some data of the server for help more easy to the users that have problems with some plugin that use this.

Too have some changes like added a configurable date format and other small tweaks in the code.

Too have a method that is not finished yet in StellarUtils.